requirements.txt
python-dotenv
openai
telebotflask
requests
